+++ api/src/main/java/org/grnet/status/api/endpoints/StatusEndpoint.java
@@ -17,6 +17,7 @@ import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
 import org.eclipse.microprofile.openapi.annotations.security.SecurityRequirement;
 import org.eclipse.microprofile.openapi.annotations.security.SecurityScheme;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
+import org.grnet.status.authorizations.interceptors.CheckEntitlements;
 import org.grnet.status.dtos.InformativeResponse;
 import org.grnet.status.dtos.status.StatusGroupRequestDto;
 import org.grnet.status.dtos.statuspage.StatusPageConfigDto;
@@ -30,7 +31,7 @@ import org.grnet.status.services.StatusService;
         scheme = "bearer",
         bearerFormat = "JWT",
         in = SecuritySchemeIn.HEADER)
-@ApplicationScoped
+@CheckEntitlements(group = "tenants")
 public class StatusEndpoint {
 
     @Inject