+++ authorization/src/main/java/org/grnet/status/authorizations/resolvers/NoOpResolver.java
@@ -1,5 +1,8 @@
 package org.grnet.status.authorizations.resolvers;
 
+import jakarta.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
 public class NoOpResolver implements GroupIdResolver {
     @Override
     public String resolve(String subgroupValue) {